w3c-reports/12092014/gtk/runned-tests/notifications/interfaces-expected.txt

CONSOLE MESSAGE: line 188: callback not yet supported
CONSOLE MESSAGE: line 188: callback not yet supported

FAIL Notification interface: existence and properties of interface object assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:966:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:947:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface object length assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1032:32
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1028:13
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: existence and properties of interface prototype object assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1073:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1071:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: existence and properties of interface prototype object's "constructor" property assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1161:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1159:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute permission assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: operation requestPermission(NotificationPermissionCallback) assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1284:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute onclick assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute onshow assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute onerror assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute onclose assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute title assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute dir assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute lang assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute body assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute tag assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute icon assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1253:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: operation close() assert_own_property: window does not have own property "Notification" expected property "Notification" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1284:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification must be primary interface of new Notification("foo") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1525:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_primary_interface_of@http://localhost:8800/resources/idlharness.js:1523:13
test_object@http://localhost:8800/resources/idlharness.js:1495:35
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Stringification of new Notification("foo") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1547:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_primary_interface_of@http://localhost:8800/resources/idlharness.js:1545:9
test_object@http://localhost:8800/resources/idlharness.js:1495:35
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "permission" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "requestPermission" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: calling requestPermission(NotificationPermissionCallback) on new Notification("foo") with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1624:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "onclick" with the proper type (2) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "onshow" with the proper type (3) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "onerror" with the proper type (4) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "onclose" with the proper type (5) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "title" with the proper type (6) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "dir" with the proper type (7) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "lang" with the proper type (8) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "body" with the proper type (9) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "tag" with the proper type (10) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "icon" with the proper type (11) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "close" with the proper type (12) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL EventTarget interface: new Notification("foo") must inherit property "addEventListener" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL EventTarget interface: calling addEventListener(DOMString,EventListener,boolean) on new Notification("foo") with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1624:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL EventTarget interface: new Notification("foo") must inherit property "removeEventListener" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL EventTarget interface: calling removeEventListener(DOMString,EventListener,boolean) on new Notification("foo") with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1624:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL EventTarget interface: new Notification("foo") must inherit property "dispatchEvent" with the proper type (2) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1583:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL EventTarget interface: calling dispatchEvent(Event) on new Notification("foo") with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Notification"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1624:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
